pyproject.toml
[project]
name = "smart-translator-agenticAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "fastmcp>=2.12.5",
    "langchain>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.0",
    "ollama>=0.6.0",
    "simple-lama-inpainting>=0.1.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pyside6>=6.10.0",
    "pillow>=9.5.0",
    "scikit-learn>=1.7.2",
    "timm>=1.0.20",
    "torch>=2.2.2",
    "torchvision>=0.24.0",
    "wandb>=0.22.2",
    "numpy>=1.24.3",
    "captum>=0.7.0",
    "transformers>=4.57.1",
    "opencv-python>=4.11.0.86",
    "gradio>=5.50.0,<6.0.0",
    "websockets<14.0",
    "optuna>=4.6.0",
    "notebook>=7.4.7",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.4.10",
    "matplotlib>=3.10.7",
    "paddleocr[all]>=3.3.2",
    "paddlepaddle>=3.2.2",
    "python-doctr>=1.0.0",
    "tqdm>=4.67.1",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
    "faiss-cpu>=1.13.1",
]


[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "optuna>=4.6.0",
    "notebook>=7.4.7",
    "jupyterlab>=4.4.10",
]
